+++ src/app/smarthome/lib/widgets/navigation.dart
@@ -1,5 +1,4 @@
 import 'package:Smarthome/constants/colors.dart';
-import 'package:curved_navigation_bar/curved_navigation_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:salomon_bottom_bar/salomon_bottom_bar.dart';
